+++ ExamService/pkg/model/exam_record.go
@@ -9,7 +9,7 @@ import (
 type ExamRecord struct {
 	Id        primitive.ObjectID `json:"_id"       bson:"_id,omitempty"`
 	ExamId    string             `json:"examId"    bson:"examId"`
-	Score     int64              `json:"score"     bson:"score"`
+	Score     int32              `json:"score"     bson:"score"`
 	UserId    string             `json:"userId"    bson:"userId"`
 	CreatedAt time.Time          `json:"createdAt" bson:"createdAt"`
 	UpdatedAt time.Time          `json:"updatedAt" bson:"updatedAt"`
